Move two unit tests from reactTest to kineticsTest
To accompany the moved methods
Also optimize imports in kineticsTest
